requirements.txt
openai>=1.0.0
boto3>=1.26.0
numpy>=1.24.0
python-dotenv>=0.19.0